inversionfix.py and biodiff should work with inversions now
inversionfix.py now outputs temp fasta with inversions undone, and a
temp vcf noting the inversions. I checked its output with 4-0010 and
the temp file has its sequence inverted compared to the copy in
$GROUPHOME/data/genomes where I expect it.

biodiff now calls dnadiff and inversionfix with appropriate arguments
for the isolate its running, and uses the temp fasta as the query.
However biodiff still cannot finish when running on 4-0010, nor can
the original biodiff.

Next I will have inversionfix.py change the transposons, hopefully
then biodiff will be able to run on 4-0010. If not, I will attempt
debugging.

Eventuallt I will write a script to take in udiff2vcf's standard
output and integrate it with the temp vcf from inversionfix.py,
and label the GAPs as insertions and deleations as appropriate.