inversionfix.py now uses blastn to refine position of inversions
This did not make the output worse, though it did not improve the
accuracy as I had hoped. The start and end positions of the
inversions are still often off by 1 bp. This may be an off-by-one
error somewhere in my code, though it is applied inconsistently.

Dnadiff is definatly off by 1 bp in its inversion detection.
I created two reference querry fasta file pairs (one based on
lambda-phage, one based on H37Rv) with 2 inversions in the same
2 locations, and ran dnadiff on both. Their positions are off
from eachother for the end position of the first inversion,
and the start and end of the 2nd inversion.