inversionfix.py now blasts querry region around inversion to reference
During the blast position refinement step, inversionfix.py was
previously blasting the region of inversion
(in terms of querry position) against querry region of inversion,
which is incorrect, as the querry position is not the equivalent
region of alingment in the reference.

Now inversionfix.py takes a region of the querry sequence around
the inversion start/stop, and blasts that against the whole
reference fasta. Then uses the first breakpoint (start of
querry alignment, + the start of the region splice in the querry)
as the new inversion start/stop.

This did not effect output for the test data, and now biodiff
can run on isolate 1-0031 without error.