inversionfix.py now refines INV positions in VCF output
inversionfix.py now calls its refineposition function
on the reference postions of inversion for VCF output
too (this BLASTs the region around the inversion start
and inversion stop in the refence to the querry sequence,
and uses the breakpoints to calculate a more accurate
position of the inversions).

inversionfix.py also uses 150 bp on eihter side of the INV
position for blasting instead of 50, to give BLAST more to
work with when refining the positions.